perf(thymeleaf): integrate Thymeleaf template engine
integrate Thymeleaf template engine
extract enV5 page handling from MonitorController
add HTML path prefix (/sensor/) for enV5 twin specific pages
update favicons

BREAKING CHANGE: Integration of Thymeleaf, Add prefix for sensor page
