+++ monitor/src/main/java/de/ude/es/ServletInitializer.java
@@ -1,0 +1,14 @@
+package de.ude.es;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(
+        SpringApplicationBuilder application
+    ) {
+        return application.sources(MonitoringServiceApplication.class);
+    }
+}
