+++ monitor/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 logging.level.web=ERROR
 server.port=8081
+server.error.include-stacktrace=always
