--- monitor/src/main/resources/html/env5.html
@@ -1,112 +1,0 @@
-<!DOCTYPE html>
-<html style="font-size: 20px;" lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="msapplication-TileColor" content="#da532c">
-    <meta name="msapplication-config" content="favicons/browserconfig.xml">
-    <meta name="theme-color" content="#ffffff">
-
-    <title>Elastic AI Monitor - ENV5</title>
-
-    <link rel="apple-touch-icon" sizes="180x180" href="favicons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="favicons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="favicons/favicon-16x16.png">
-    <link rel="manifest" href="favicons/site.webmanifest">
-    <link rel="mask-icon" href="favicons/safari-pinned-tab.svg" color="#5bbad5">
-    <link rel="shortcut icon" href="favicons/favicon.ico">
-
-    <!-- Bootstrap -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
-
-    <!-- Custom JS -->
-    <script>
-        function getRootUrl() {
-            return window.location.origin
-                ? window.location.origin + "/"
-                : window.location.protocol + "/" + window.location.host + "/";
-        }
-
-        async function getValueUpdate(deviceId, sensorId, fieldId) {
-            const response = await fetch(getRootUrl() + deviceId + "/" + sensorId);
-
-            if (response.status !== 200) {
-                console.log(response.json());
-                document.getElementById(fieldId).value = "HTTP Status: " + response.status;
-                return;
-            }
-
-            const result = await response.json();
-            document.getElementById(fieldId).value = result.value;
-        }
-    </script>
-</head>
-<body>
-<section class="d-flex flex-column align-items-center pt-1">
-    <h1>Elastic AI Monitor</h1>
-    <h3>TWIN_NAME</h3>
-    <div class="row">
-        <div class="col">
-            <button type="button" class="btn btn-primary" onClick="window.location.reload();">Refresh</button>
-            <a class="btn btn-primary" href="/index">Home</a>
-        </div>
-    </div>
-</section>
-<section class="d-flex flex-column align-items-center pt-3">
-    <h4>Twin Data</h4>
-    <div class="container">
-        <table class="table table-borderless table-hover">
-            <tbody>
-            <tr>
-                <th class="text-end pe-2" scope="row">Name</th>
-                <td>TWIN_NAME</td>
-            </tr>
-            <tr>
-                <th class="text-end pe-2" scope="row">ID</th>
-                <td>TWIN_ID</td>
-            </tr>
-            <tr>
-                <th class="text-end pe-2" scope="row">Status</th>
-                <td>TWIN_STATUS</td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-</section>
-<section class="d-flex flex-column align-items-center pt-3">
-    <h4>Measurements</h4>
-    <div class="d-grid container gap-2">
-        <div class="row text-center">
-            <h5>Sensor 1</h5>
-            <div class="input-group">
-                <span class="input-group-text" id="label-wifi">WIFI</span>
-                <input type="text" class="form-control" aria-label="wifi" aria-describedby="label-wifi" id="wifi"
-                       readonly>
-            </div>
-            <script>
-                getValueUpdate("TWIN_ID", "wifi", "wifi");
-                setInterval("getValueUpdate(\"TWIN_ID\", \"wifi\", \"wifi\");", 10000);
-            </script>
-        </div>
-        <div class="row text-center">
-            <h5>Sensor 2</h5>
-            <div class="input-group">
-                <span class="input-group-text" id="label-sram">SRAM</span>
-                <input type="text" class="form-control" aria-label="sram" aria-describedby="label-sram" id="sram"
-                       readonly>
-            </div>
-            <script>
-                getValueUpdate("TWIN_ID", "sram", "sram");
-                setInterval("getValueUpdate(\"TWIN_ID\", \"sram\", \"sram\");", 10000);
-            </script>
-        </div>
-    </div>
-</section>
-<section class="d-flex flex-column align-items-center pt-3">
-    <h4>Flash FPGA</h4>
-    <div class="text-secondary">
-        WORK IN PROGRESS
-    </div>
-</section>
-</body>
-</html>
