--- monitor/src/main/resources/public/error/404.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html style="font-size: 20px;" lang="en">
-<head>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-            crossorigin="anonymous"></script>
-
-    <title>Error 404</title>
-</head>
-<body>
-<div class="position-absolute top-50 start-50 translate-middle">
-    <div class="container">
-        <div class="row text-center">
-            <h1 class="display-3">Error 404</h1>
-        </div>
-        <div class="row text-center">
-            <p>Maybe this page moved? Got deleted? Is hiding out in quarantine? Never existed in the first place?</p>
-            <p>Let's go <a href="http://localhost:8081/">home</a> and try from there.</p>
-        </div>
-    </div>
-</div>
-</body>
-</html>
